+++ mitmproxy/proxy/context.py
@@ -38,7 +38,8 @@ class Context:
         self.client = client
         self.options = options
         self.server = connection.Server(
-            None, transport_protocol=client.transport_protocol
+            address=None,
+            transport_protocol=client.transport_protocol
         )
         self.layers = []