+++ test/mitmproxy/coretypes/test_serializable.py
@@ -1,6 +1,16 @@
+from __future__ import annotations
+
 import copy
+import dataclasses
+import enum
+from collections.abc import Mapping
+from dataclasses import dataclass
+from typing import Literal, Optional
+
+import pytest
 
 from mitmproxy.coretypes import serializable
+from mitmproxy.coretypes.serializable import SerializableDataclass
 
 
 class SerializableDummy(serializable.Serializable):
@@ -34,3 +44,130 @@ class TestSerializable:
         b = a.copy()
         assert a.get_state()["id"] != b.get_state()["id"]
         assert a.get_state()["foo"] == b.get_state()["foo"]
+
+
+@dataclass
+class Simple(SerializableDataclass):
+    x: int
+    y: Optional[str]
+
+
+@dataclass
+class SerializableChild(SerializableDataclass):
+    foo: Simple
+    maybe_foo: Optional[Simple]
+
+
+@dataclass
+class Inheritance(Simple):
+    z: bool
+
+
+class TEnum(enum.Enum):
+    A = 1
+    B = 2
+
+
+@dataclass
+class TLiteral(SerializableDataclass):
+    l: Literal["foo", "bar"]
+
+
+@dataclass
+class BuiltinChildren(SerializableDataclass):
+    a: Optional[list[int]]
+    b: Optional[dict[str, int]]
+    c: Optional[tuple[int, int]]
+    d: list[Simple]
+    e: Optional[TEnum]
+
+
+@dataclass
+class Defaults(SerializableDataclass):
+    z: Optional[int] = 42
+
+
+@dataclass
+class Unsupported(SerializableDataclass):
+    a: Mapping[str, int]
+
+
+@dataclass
+class Addr(SerializableDataclass):
+    peername: tuple[str, int]
+
+
+@dataclass(frozen=True)
+class Frozen(SerializableDataclass):
+    x: int
+
+
+@dataclass
+class FrozenWrapper(SerializableDataclass):
+    f: Frozen
+
+
+class TestSerializableDataclass:
+    @pytest.mark.parametrize("cls, state", [
+        (Simple, {"x": 42, "y": 'foo'}),
+        (Simple, {"x": 42, "y": None}),
+        (SerializableChild, {"foo": {"x": 42, "y": "foo"}, "maybe_foo": None}),
+        (SerializableChild, {"foo": {"x": 42, "y": "foo"}, "maybe_foo": {"x": 42, "y": "foo"}}),
+        (Inheritance, {"x": 42, "y": "foo", "z": True}),
+        (BuiltinChildren, {"a": [1, 2, 3], "b": {"foo": 42}, "c": (1, 2), "d": [{"x": 42, "y": "foo"}], "e": 1}),
+        (BuiltinChildren, {"a": None, "b": None, "c": None, "d": [], "e": None}),
+        (TLiteral, {"l": "foo"}),
+    ])
+    def test_roundtrip(self, cls, state):
+        a = cls.from_state(copy.deepcopy(state))
+        assert a.get_state() == state
+
+    def test_set(self):
+        s = SerializableChild(foo=Simple(x=42, y=None), maybe_foo=Simple(x=43, y=None))
+        s.set_state({"foo": {"x": 44, "y": None}, "maybe_foo": None})
+        assert s.foo.x == 44
+        assert s.maybe_foo is None
+        with pytest.raises(ValueError, match="Unexpected fields"):
+            Simple(0, "").set_state({"x": 42, "y": "foo", "z": True})
+
+    def test_invalid_none(self):
+        with pytest.raises(ValueError):
+            Simple.from_state({"x": None, "y": "foo"})
+
+    def test_defaults(self):
+        a = Defaults()
+        assert a.get_state() == {"z": 42}
+
+    def test_invalid_type(self):
+        with pytest.raises(ValueError):
+            Simple.from_state({"x": 42, "y": 42})
+        with pytest.raises(ValueError):
+            BuiltinChildren.from_state({"a": None, "b": None, "c": ("foo",), "d": [], "e": None})
+
+    def test_invalid_key(self):
+        with pytest.raises(ValueError):
+            Simple.from_state({"x": 42, "y": "foo", "z": True})
+
+    def test_invalid_type_in_list(self):
+        with pytest.raises(ValueError, match="Invalid value for x"):
+            BuiltinChildren.from_state({"a": None, "b": None, "c": None, "d": [{"x": "foo", "y": "foo"}], "e": None})
+
+    def test_unsupported_type(self):
+        with pytest.raises(TypeError):
+            Unsupported.from_state({"a": "foo"})
+
+    def test_literal(self):
+        assert TLiteral.from_state({"l": "foo"}).get_state() == {"l": "foo"}
+        with pytest.raises(ValueError):
+            TLiteral.from_state({"l": "unknown"})
+
+    def test_peername(self):
+        assert Addr.from_state({"peername": ("addr", 42)}).get_state() == {"peername": ("addr", 42)}
+        assert Addr.from_state({"peername": ("addr", 42, 0, 0)}).get_state() == {"peername": ("addr", 42, 0, 0)}
+
+    def test_set_immutable(self):
+        w = FrozenWrapper(Frozen(42))
+        with pytest.raises(dataclasses.FrozenInstanceError):
+            w.f.set_state({"x": 43})
+        w.set_state({"f": {"x": 43}})
+        assert w.f.x == 43